# pref(gatsby-plugin-mdx): add lessBabel option, 40% perf win