--- a/e2e-tests/mdx-less-babel/.gitignore
+++ b/e2e-tests/mdx-less-babel/.gitignore
@@ -1,0 +1,11 @@
+# Project dependencies
+.cache
+node_modules
+yarn-error.log
+
+# Build directory
+/public
+.DS_Store
+
+# Cypress output
+cypress/videos/