--- a/e2e-tests/mdx-less-babel/cypress.json
+++ b/e2e-tests/mdx-less-babel/cypress.json
@@ -1,0 +1,3 @@
+{
+  "baseUrl": "http://localhost:9000"
+}