--- a/e2e-tests/mdx-less-babel/gatsby-config.js
+++ b/e2e-tests/mdx-less-babel/gatsby-config.js
@@ -20,7 +20,7 @@ module.exports = {
     {
       resolve: `gatsby-plugin-mdx`,
       options: {
-        lessBabel: false,
+        lessBabel: true,
         extensions: [`.mdx`, `.md`],
         defaultLayouts: {
           default: require.resolve("./src/components/layout.js"),