--- a/e2e-tests/mdx-less-babel/package.json
+++ b/e2e-tests/mdx-less-babel/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "gatsby-mdx-tester",
+  "description": "Gatsby MDX tester",
+  "version": "1.0.0",
+  "dependencies": {
+    "@mdx-js/mdx": "^1.6.6",
+    "@mdx-js/react": "^1.6.6",
+    "cypress": "^3.1.0",
+    "gatsby": "^2.0.118",
+    "gatsby-plugin-mdx": "^1.2.19",
+    "gatsby-source-filesystem": "^2.3.14",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
+    "theme-ui": "^0.3.1"
+  },
+  "keywords": [
+    "gatsby"
+  ],
+  "license": "MIT",
+  "scripts": {
+    "build": "gatsby build",
+    "develop": "gatsby develop",
+    "format": "prettier --write '**/*.js'",
+    "test": "cross-env CYPRESS_SUPPORT=y npm run build && npm run start-server-and-test",
+    "start-server-and-test": "start-server-and-test serve http://localhost:9000 cy:run",
+    "serve": "gatsby serve",
+    "cy:open": "cypress open",
+    "cy:run": "(is-ci && cypress run --browser chrome --record) || cypress run --browser chrome"
+  },
+  "devDependencies": {
+    "cross-env": "^5.2.0",
+    "gatsby-cypress": "^0.1.7",
+    "is-ci": "^2.0.0",
+    "prettier": "2.0.4",
+    "start-server-and-test": "^1.7.1"
+  }
+}