--- a/e2e-tests/mdx-less-babel/src/components/example.js
+++ b/e2e-tests/mdx-less-babel/src/components/example.js
@@ -1,0 +1,7 @@
+import React from "react"
+
+const Example = () => (
+    <div data-testid="shortcode">I am an example of a component in MDX.</div>
+)
+
+export default Example