--- a/e2e-tests/mdx-less-babel/src/components/layout.js
+++ b/e2e-tests/mdx-less-babel/src/components/layout.js
@@ -1,0 +1,10 @@
+import React from "react"
+import { MDXProvider } from "@mdx-js/react"
+import Example from "./example"
+
+const shortcodes = { Example }
+export default function Layout({ children }) {
+    return (
+        <MDXProvider components={shortcodes}>{children}</MDXProvider>
+    )
+}