--- a/e2e-tests/mdx-less-babel/src/pages/another.mdx
+++ b/e2e-tests/mdx-less-babel/src/pages/another.mdx
@@ -1,0 +1,1 @@
+This is to add another source page for slugification.