--- a/e2e-tests/mdx-less-babel/src/pages/index.mdx
+++ b/e2e-tests/mdx-less-babel/src/pages/index.mdx
@@ -1,0 +1,17 @@
+import { Message } from "theme-ui"
+
+Let's go MDX!
+
+## Do you work
+
+- nobody knows
+
+````
+```javascript
+const codefence = true
+```
+````
+
+<Example/>
+
+<Message data-testid="external">Now an external import</Message>