--- a/e2e-tests/mdx-less-babel/src/posts/about/embedded.mdx
+++ b/e2e-tests/mdx-less-babel/src/posts/about/embedded.mdx
@@ -1,0 +1,1 @@
+This is a page that should include a slash slug