--- a/e2e-tests/mdx-less-babel/src/posts/about/index.mdx
+++ b/e2e-tests/mdx-less-babel/src/posts/about/index.mdx
@@ -1,0 +1,1 @@
+This should be an about page we can access with a nested slug.