--- a/e2e-tests/mdx-less-babel/src/posts/my-blog.md
+++ b/e2e-tests/mdx-less-babel/src/posts/my-blog.md
@@ -1,0 +1,1 @@
+Let's see if we can generate a slug for this too