--- a/packages/gatsby-plugin-mdx/gatsby/on-create-node.js
+++ b/packages/gatsby-plugin-mdx/gatsby/on-create-node.js
@@ -180,7 +180,6 @@ export default { ${scopeIdentifiers.join(`, `)} }`
       configFile: false,
       plugins: [instance.plugin],
     })
-
     scopeFileContent = result.code
   }