--- a/packages/gatsby-plugin-mdx/loaders/mdx-loader.js
+++ b/packages/gatsby-plugin-mdx/loaders/mdx-loader.js
@@ -25,7 +25,10 @@ const debugMore = require(`debug`)(`gatsby-plugin-mdx-info:mdx-loader`)
 
 const genMdx = require(`../utils/gen-mdx`)
 const withDefaultOptions = require(`../utils/default-options`)
-const createMDXNode = require(`../utils/create-mdx-node`)
+const {
+  createMdxNodeExtraBabel,
+  createMdxNodeLessBabel,
+} = require(`../utils/create-mdx-node`)
 const { createFileNode } = require(`../utils/create-fake-file-node`)
 
 const DEFAULT_OPTIONS = {
@@ -89,11 +92,12 @@ const hasDefaultExport = (str, options) => {
   return hasDefaultExportBool
 }
 
-module.exports = async function (content) {
+module.exports = async function mdxLoader(content) {
   const callback = this.async()
   const {
     getNode: rawGetNode,
     getNodes,
+    getNodesByType,
     reporter,
     cache,
     pathPrefix,
@@ -121,11 +125,23 @@ module.exports = async function (content) {
 
   let mdxNode
   try {
-    mdxNode = await createMDXNode({
-      id: `fakeNodeIdMDXFileABugIfYouSeeThis`,
-      node: fileNode,
-      content,
-    })
+    // This fakeNodeIdMDXFileABugIfYouSeeThis node object attempts to break the chicken-egg problem,
+    // where parsing mdx allows for custom plugins, which can receive a mdx node.
+    if (options.lessBabel) {
+      ;({ mdxNode } = await createMdxNodeLessBabel({
+        id: `fakeNodeIdMDXFileABugIfYouSeeThis`,
+        node: fileNode,
+        content,
+        options,
+        getNodesByType,
+      }))
+    } else {
+      mdxNode = await createMdxNodeExtraBabel({
+        id: `fakeNodeIdMDXFileABugIfYouSeeThis`,
+        node: fileNode,
+        content,
+      })
+    }
   } catch (e) {
     return callback(e)
   }
@@ -192,6 +208,7 @@ ${contentWithoutFrontmatter}`
     node: { ...mdxNode, rawBody: code },
     getNode,
     getNodes,
+    getNodesByType,
     reporter,
     cache,
     pathPrefix,