--- a/packages/gatsby-plugin-mdx/utils/create-mdx-node.js
+++ b/packages/gatsby-plugin-mdx/utils/create-mdx-node.js
@@ -3,7 +3,9 @@ const { createContentDigest } = require(`gatsby-core-utils`)
 const mdx = require(`../utils/mdx`)
 const extractExports = require(`../utils/extract-exports`)
 
-module.exports = async ({ id, node, content }) => {
+const { findImportsExports } = require(`../utils/gen-mdx`)
+
+async function createMdxNodeExtraBabel({ id, node, content }) {
   let code
   try {
     code = await mdx(content)
@@ -47,3 +49,50 @@ module.exports = async ({ id, node, content }) => {
 
   return mdxNode
 }
+
+async function createMdxNodeLessBabel({ id, node, content, ...helpers }) {
+  const {
+    frontmatter,
+    scopeImports,
+    scopeExports,
+    scopeIdentifiers,
+  } = await findImportsExports({
+    mdxNode: node,
+    rawInput: content,
+    absolutePath: node.absolutePath,
+    ...helpers,
+  })
+
+  const mdxNode = {
+    id,
+    children: [],
+    parent: node.id,
+    internal: {
+      content: content,
+      type: `Mdx`,
+    },
+  }
+
+  mdxNode.frontmatter = {
+    title: ``, // always include a title
+    ...frontmatter,
+  }
+
+  mdxNode.excerpt = frontmatter.excerpt
+  mdxNode.exports = scopeExports
+  mdxNode.rawBody = content
+
+  // Add path to the markdown file path
+  if (node.internal.type === `File`) {
+    mdxNode.fileAbsolutePath = node.absolutePath
+  }
+
+  mdxNode.internal.contentDigest = createContentDigest(mdxNode)
+
+  return { mdxNode, scopeIdentifiers, scopeImports }
+}
+
+module.exports = createMdxNodeExtraBabel // Legacy default export
+
+module.exports.createMdxNodeExtraBabel = createMdxNodeExtraBabel
+module.exports.createMdxNodeLessBabel = createMdxNodeLessBabel