… packages/gatsby-plugin-mdx/loaders/__snapshots__/mdx-loader.test.js.snap