@mxstbr 2020-11-10
> Once we get more feedback we can bump the major and flip the switch to make it opt-out instead. But that's future.

Love it! 👍 Thanks for tackling this :100: